Move builtin types to the module defining the corresponding type
