--- packages/kryo/src/lib/builtins/boolean.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/boolean
- */
-
-import { BooleanType } from "../boolean.js";
-
-export const $Boolean: BooleanType = new BooleanType();
