--- packages/kryo/src/lib/builtins/date.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/date
- */
-
-import { DateType } from "../date.js";
-
-export const $Date: DateType = new DateType();
