--- packages/kryo/src/lib/builtins/null.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/null
- */
-
-import { NullType } from "../null.js";
-
-export const $Null: NullType = new NullType();
