--- packages/kryo/src/lib/builtins/sint54.ts
@@ -1,10 +1,0 @@
-/**
- * @module kryo/builtins/sint54
- */
-
-import { IntegerType } from "../integer.js";
-
-export const $Sint54: IntegerType = new IntegerType({
-  min: -9007199254740992,
-  max: 9007199254740991,
-});
