--- packages/kryo/src/lib/builtins/sint8.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/sint8
- */
-
-import { IntegerType } from "../integer.js";
-
-export const $Sint8: IntegerType = new IntegerType({min: -128, max: 127});
