--- packages/kryo/src/lib/builtins/ucs2-string.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/ucs2-string
- */
-
-import { Ucs2StringType } from "../ucs2-string.js";
-
-export const $Ucs2String: Ucs2StringType = new Ucs2StringType({maxLength: Infinity});
