--- packages/kryo/src/lib/builtins/uint16.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/uint16
- */
-
-import { IntegerType } from "../integer.js";
-
-export const $Uint16: IntegerType = new IntegerType({min: 0, max: 65535});
