--- packages/kryo/src/lib/builtins/uint53.ts
@@ -1,7 +1,0 @@
-/**
- * @module kryo/builtins/uint53
- */
-
-import { IntegerType } from "../integer.js";
-
-export const $Uint53: IntegerType = new IntegerType({min: 0, max: 9007199254740991});
